+++ RetrospectiveExtension.Frontend/components/__tests__/feedbackCarousel.test.tsx
@@ -3,6 +3,7 @@ import { shallow } from 'enzyme';
 import FeedbackCarousel from '../../components/feedbackCarousel';
 import FeedbackItem from '../../components/feedbackItem';
 import { testGroupColumnProps, testColumnProps, testGroupFeedbackItemTwo } from '../__mocks__/mocked_components/mockedFeedbackColumn';
+import { mockUuid } from '../__mocks__/uuid/v4';
 
 const mockedProps = {
   feedbackColumnPropsList: [testColumnProps],
@@ -15,6 +16,8 @@ const mockedGroupProps = {
   isFeedbackAnonymous: true,
   isFocusModalHidden: false
 };
+
+jest.mock('uuid', () => ({ v4: () => mockUuid}));
 
 describe('Feedback Carousel ', () => {
   it('can be rendered', () => {
@@ -40,5 +43,24 @@ describe('Feedback Carousel ', () => {
     const feedbackItem = component.findWhere(c => c.prop('className') === 'feedback-carousel-item').find(FeedbackItem);
 
     expect(feedbackItem.prop('groupIds')).toEqual([testGroupFeedbackItemTwo.id]);
+
+  describe("'All' column", () => {
+    it("should be set by default in the first position", () => {
+      const wrapper = shallow(<FeedbackCarousel {...mockedProps} />);
+      const component = wrapper.children().dive();
+  
+      const allColumn = component.findWhere(c => c.prop('headerText')).first();
+  
+      expect(allColumn.prop('headerText')).toEqual('All');
+    });
+
+    it("should not exist when there are no feedback columns", () => {
+      const wrapper = shallow(<FeedbackCarousel feedbackColumnPropsList={[]} isFeedbackAnonymous={true} isFocusModalHidden={false} />);
+      const component = wrapper.children().dive();
+  
+      const allColumn = component.findWhere(c => c.prop('headerText')).first();
+  
+      expect(allColumn).toHaveLength(0);
+    });
   })
 });