+++ CHANGELOG.md
@@ -10,6 +10,7 @@ _PS: Unfortunately, changelog before v1.0.46 is not available_ 🤦‍♂️
 closely resemble the Retrospective summary's color separation for the three categories:
 Reds and Oranges for Unfavorable (1-6), Yellows for Neutral (7-8), Greens for
 Favorable (9-10). From [Github PR #531](https://github.com/microsoft/vsts-extension-retrospectives/pull/531).
+* Related feedback items, in "Focus Mode", now show the original column textual as well as visually. From [Github PR #544](https://github.com/microsoft/vsts-extension-retrospectives/pull/544)
 
 ## v1.91.1