+++ CHANGELOG.md
@@ -10,6 +10,8 @@ _PS: Unfortunately, changelog before v1.0.46 is not available_ 🤦‍♂️
 closely resemble the Retrospective summary's color separation for the three categories:
 Reds and Oranges for Unfavorable (1-6), Yellows for Neutral (7-8), Greens for
 Favorable (9-10). From [Github PR #531](https://github.com/microsoft/vsts-extension-retrospectives/pull/531).
+* New tab in "Focus Mode", called "All", which contains every card on the current retrospective board so that
+your team can prioritize the highest voted cards first. From [Github PR #531](https://github.com/microsoft/vsts-extension-retrospectives/pull/543).
 
 ## v1.91.1