+++ CHANGELOG.md
@@ -4,7 +4,7 @@ You can find the changelog of the Retrospective Extension below.
 
 _PS: Unfortunately, changelog before v1.0.46 is not available_ 🤦‍♂️
 
-## v1.XX.XX
+## v1.92.1
 
 * Team Assessment form: Background colors for each number on the spectrum now more
 closely resemble the Retrospective summary's color separation for the three categories: